requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn
pydantic==2.5.0
pydantic-settings==2.1.0
# Azure Services
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
azure-monitor-opentelemetry==1.1.0
openai==1.3.7

# Audio Processing
openai-whisper==20231117
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# Database
sqlalchemy==2.0.23
pyodbc==5.0.1
alembic==1.13.0

# Security & Encryption
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Authentication
python-jose==3.3.0
pyjwt==2.8.0

# Logging & Monitoring
python-json-logger==2.0.7
structlog==23.2.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
tenacity==8.2.3
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.1.0

# Code Quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.2
pydantic[email]
bcrypt==4.1.2
pymssql